agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.2",
    "python-dotenv>=1.1.0",
    "langchain-openai>=0.0.2",
    "pydantic>=2.0.0",
    "fastapi>=0.115.12",
]
